requirements.txt
fastapi
uvicorn
pandas
python-dotenv
typer
kiteconnect
requests
google-genai
upstox-python-sdk
groq
SQLAlchemy
alembic
pytest
email-validator
cryptography
python-jose[cryptography]
passlib[bcrypt]
python-multipart
